+++ packages/orq-rc/RELEASES.md
@@ -11189,3 +11189,13 @@ Based on:
 - [python v4.16.0-rc.17] packages/orq-rc
 ### Releases
 - [PyPI v4.16.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.17 - packages/orq-rc
+
+## 2026-09-25 07:44:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.799.0 (2.941.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.16.0-rc.18] packages/orq-rc
+### Releases
+- [PyPI v4.16.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.18 - packages/orq-rc